update typhoeus instrumentation and add prepend
