+++ lib/new_relic/agent/configuration/default_source.rb
@@ -890,6 +890,14 @@ module NewRelic
 
           :description => "Controls auto-instrumentation of Net::HTTP at start up.  May be one of [auto|prepend|chain|disabled]."
         },
+        :'instrumentation.typhoeus' => {
+          :default => instrumentation_value_of(:disable_typhoeus),
+          :public => :true,
+          :type => String,
+          :dynamic_name => true,
+          :allowed_from_server => false,
+          :description => "Controls auto-instrumentation of Typhoeus at start up.  May be one of [auto|prepend|chain|disabled]."
+        },
         :disable_data_mapper => {
           :default => false,
           :public => true,
@@ -1554,7 +1562,7 @@ module NewRelic
           :type         => Boolean,
           :dynamic_name => true,
           :allowed_from_server => false,
-          :description  => 'If <code>true</code>, the agent won\'t install instrumentation for the typhoeus gem.'
+          :description  => deprecated_description(:'instrumentation.typhoeus', 'If <code>true</code>, the agent won\'t install instrumentation for the typhoeus gem.' )
         },
         :disable_httprb => {
           :default      => false,
